+++ src/main/java/duke/commands/Command.java
@@ -1,3 +1,7 @@
+package duke.commands;
+import duke.utils.Storage;
+import duke.tasks.*;
+
 import java.util.ArrayList;
 
 public abstract class Command {
