+++ src/main/java/duke/commands/DeleteCommand.java
@@ -1,18 +1,23 @@
+package duke.commands;
+
+import duke.utils.*;
+import duke.tasks.*;
 import java.util.ArrayList;
 
-public class MarkCommand extends Command {
+public class DeleteCommand extends Command {
     private int taskNumber;
 
-    public MarkCommand(int taskNumber) {
+    public DeleteCommand(int taskNumber) {
         this.taskNumber = taskNumber;
     }
 
     @Override
     public void execute(Storage storage, ArrayList<Task> tasks) {
         try {
-            tasks.get(taskNumber - 1).markDone();
-            System.out.println("Nice! I've marked this task as done:");
-            System.out.println(tasks.get(taskNumber - 1));
+            Task task = tasks.get(taskNumber - 1);
+            System.out.println("Noted. I've removed this task:");
+            System.out.println(task);
+            tasks.remove(this.taskNumber-1);
             Ui.printLine();
         } catch (IndexOutOfBoundsException | NumberFormatException e) {
             System.out.println("The selected task does not exist");
