+++ src/main/java/duke/commands/ExitCommand.java
@@ -1,3 +1,7 @@
+package duke.commands;
+
+import duke.utils.*;
+import duke.tasks.*;
 import java.io.IOException;
 import java.util.ArrayList;
