+++ src/main/java/duke/commands/MarkCommand.java
@@ -1,19 +1,22 @@
+package duke.commands;
+
+import duke.utils.*;
+import duke.tasks.*;
 import java.util.ArrayList;
 
-public class DeleteCommand extends Command {
+public class MarkCommand extends Command {
     private int taskNumber;
 
-    public DeleteCommand(int taskNumber) {
+    public MarkCommand(int taskNumber) {
         this.taskNumber = taskNumber;
     }
 
     @Override
     public void execute(Storage storage, ArrayList<Task> tasks) {
         try {
-            Task task = tasks.get(taskNumber - 1);
-            System.out.println("Noted. I've removed this task:");
-            System.out.println(task);
-            tasks.remove(this.taskNumber-1);
+            tasks.get(taskNumber - 1).markDone();
+            System.out.println("Nice! I've marked this task as done:");
+            System.out.println(tasks.get(taskNumber - 1));
             Ui.printLine();
         } catch (IndexOutOfBoundsException | NumberFormatException e) {
             System.out.println("The selected task does not exist");
