+++ src/main/java/duke/tasks/ToDo.java
@@ -1,3 +1,5 @@
+package duke.tasks;
+
 public class ToDo extends Task {
     public ToDo(String description) {
         super(description);
