+++ src/main/java/duke/utils/DukeException.java
@@ -1,3 +1,5 @@
+package duke.utils;
+
 public class DukeException extends Exception {
     public DukeException(String message) {
         super(message);
