+++ src/main/java/duke/utils/Storage.java
@@ -1,9 +1,12 @@
+package duke.utils;
+
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileWriter;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Scanner;
+import duke.tasks.*;
 
 public class Storage {
     private String filepath;
